requirements.txt
Pillow==9.0.1
requests==2.28.1
numpy==1.23.1
torch==1.13.1